pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cobol-legacy-ledger"
version = "6.1.0"
description = "Educational COBOL banking system with Python integrity layer, REST API, LLM tool-use, legacy payroll sidecar, and static analysis pipeline"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "Albert Mendes" },
]
keywords = ["cobol", "banking", "education", "integrity", "settlement", "legacy", "analysis", "payroll", "spaghetti-code", "static-analysis"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Education",
]
dependencies = [
    "click>=8.1.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.26.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.43.0"]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
]

[project.scripts]
ledger = "python.cli:cli"

[project.urls]
Repository = "https://github.com/gitgoodordietrying/cobol-legacy-ledger"
Documentation = "https://github.com/gitgoodordietrying/cobol-legacy-ledger/tree/main/docs"
Issues = "https://github.com/gitgoodordietrying/cobol-legacy-ledger/issues"

[tool.setuptools.packages.find]
include = ["python*"]

[tool.pytest.ini_options]
testpaths = ["python/tests"]
asyncio_mode = "auto"
